Some renames and additional typing
